Containerised the program

+++ src/backup.py
@@ -7,7 +7,7 @@ import datetime
 # Get the backup folder path from environment variable
 backup_folder_path = os.environ.get('BACKUPFOLDER')
 if backup_folder_path is None or not backup_folder_path:
-    backup_folder_path = './backups'
+    backup_folder_path = './FCA-Flask/backups'
 
 # Create a backup folder if it doesn't exist
 if not os.path.exists(backup_folder_path):
@@ -21,10 +21,15 @@ def folder(source_folder_path):
 
     # Create a timestamp for the backup folder name
     timestamp = datetime.datetime.now().strftime("%Y%m%d%H%M%S")
-    backup_folder_name = f"{os.path.basename(source_folder_path)}_{timestamp}"
-    backup_path = os.path.join(backup_folder_path, backup_folder_name)
 
     # Copy the contents of the source folder to the backup folder
-    shutil.copytree(source_folder_path, backup_path)
+    if os.path.isfile(source_folder_path):
+        backup_folder_name = f"{timestamp}_{os.path.basename(source_folder_path)}"
+        backup_path = os.path.join(backup_folder_path, backup_folder_name)
+        shutil.copyfile(source_folder_path, backup_path)
+    else:
+        backup_file_name = f"{os.path.basename(source_folder_path)}_{timestamp}"
+        backup_path = os.path.join(backup_folder_path, backup_file_name)
+        shutil.copytree(source_folder_path, backup_path)
 
     log_to_db("BACKUP", "Folder=" + source_folder_path, "SUCCESS")
